requirements.txt
requests>=2.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
PyYAML>=6.0
pytest>=7.2.0
python-dateutil>=2.8.2
